Conflicts:
	CCTM/src/cio/centralized_io_module.F
	CCTM/src/emis/emis/EMIS_DEFN.F
	CCTM/src/emis/emis/SSEMIS.F
	CCTM/src/emis/emis/em_param_module.F

+++ CCTM/scripts/bldit_cctm.csh
@@ -115,21 +115,25 @@ set make_options = "-j"                #> additional options for make command if
  set ModPhot   = phot/inline                #> photolysis calculation module 
                                             #>     (see $CMAQ_MODEL/CCTM/src/phot)
 
- setenv Mechanism cb6r3_ae7_aq               #> chemical mechanism (see $CMAQ_MODEL/CCTM/src/MECHS)
+ setenv Mechanism cb6r5_ae7_aq              #> chemical mechanism (see $CMAQ_MODEL/CCTM/src/MECHS)
+
  set ModMech   = MECHS/${Mechanism}
  
- set ChemSolver = ebi                       #> gas-phase chemistry solver (see $CMAQ_MODEL/CCTM/src/gas)
-                                            #> use gas/ros3 or gas/smvgear for a solver independent 
-                                            #> of the photochemical mechanism [ default: ebi ]
+ if ( ${Mechanism} != cb6r5m_ae7_aq ) then  #> Gas-phase chemistry solver options
+     set ChemSolver = ebi                   #> gas-phase chemistry solver (see $CMAQ_MODEL/CCTM/src/gas)
+ else                                       #> use gas/ros3 or gas/smvgear for a solver independent
+     set ChemSolver = ros3                  #> of the photochemical mechanism [ default for most mechanisms: ebi ]
+ endif
+                                         
  if ( $ChemSolver == ebi ) then             
-    set ModGas    = gas/${ChemSolver}_${Mechanism}   
- else                                       
+    set ModGas    = gas/${ChemSolver}_${Mechanism}
+                                            
+ else
     set ModGas    = gas/${ChemSolver}
  endif
     
  set ModAero   = aero/aero7                 #> aerosol chemistry module (see $CMAQ_MODEL/CCTM/src/aero)
  set ModCloud  = cloud/acm_ae7              #> cloud chemistry module (see $CMAQ_MODEL/CCTM/src/cloud)
-                                            #>   overwritten below if using cb6r3m_ae7_kmtbr mechanism
  set ModUtil   = util/util                  #> CCTM utility modules
  set ModDiag   = diag                       #> CCTM diagnostic modules
  set Tracer    = trac0                      #> tracer configuration directory under 
@@ -296,11 +300,6 @@ set make_options = "-j"                #> additional options for make command if
       echo "bldit_mech did not finish correctly --> Build Process Halted"
       exit 1
     endif
- endif
-
-#> Cloud chemistry options
- if ( ${Mechanism} == cb6r3m_ae7_kmtbr ) then
-    set ModCloud = cloud/acm_ae7_kmtbr
  endif
 
 #> Tracer configuration files
@@ -585,7 +584,8 @@ set Cfile = ${Bld}/${CFG}.bld      # Config Filename
  echo "Module ${ModGas};"                                          >> $Cfile
  echo                                                              >> $Cfile
 
- set MechList = " cb6mp_ae6_aq, cb6r3_ae6_aq, cb6r3_ae7_aq, cb6r3_ae7_aqkmt2, cb6r3m_ae7_kmtbr, racm2_ae6_aq, saprc07tc_ae6_aq, saprc07tic_ae6i_aq, saprc07tic_ae6i_aqkmti, saprc07tic_ae7i_aq, saprc07tic_ae7i_aqkmt2"
+ set MechList = "cb6r5hap_ae7_aq, cb6r3_ae7_aq, cb6r5_ae7_aq, cb6r5_ae7_aqkmt2, cb6r5m_ae7_aq, racm2_ae6_aq, saprc07tc_ae6_aq, saprc07tic_ae6i_aq, saprc07tic_ae6i_aqkmti, saprc07tic_ae7i_aq, saprc07tic_ae7i_aqkmt2"
+
  set text = "gas chemistry mechanisms"
  echo "// " $text                                                  >> $Cfile
  set text = "$MechList"
@@ -703,13 +703,19 @@ set Cfile = ${Bld}/${CFG}.bld      # Config Filename
     set bld_flags = "${bld_flags} -isam_cctm"
  endif
 
+ if ( $?build_twoway ) then
+   set bld_flags = "${bld_flags} -twoway"
+ endif
+
 #> Run BLDMAKE with source code in build directory
  $Blder $bld_flags $Cfile   
 
 #> Rename Makefile to specify compiler option and link back to Makefile
- mv Makefile Makefile.$compilerString
- if ( -e Makefile.$compilerString && -e Makefile ) rm Makefile
- ln -s Makefile.$compilerString Makefile
+ if ( ! $?build_twoway ) then
+    mv Makefile Makefile.$compilerString
+    if ( -e Makefile.$compilerString && -e Makefile ) rm Makefile
+    ln -s Makefile.$compilerString Makefile
+ endif
 
 #> Alert user of error in BLDMAKE if it ocurred
  if ( $status != 0 ) then
